TimeDataのDeltaTimeをdoubleに変更した

+++ Scripts/Core/TimeData.cs
@@ -3,6 +3,6 @@ namespace mm.core
     [System.Serializable]
     public struct TimeData
     {
-        public float DeltaTime;
+        public double DeltaTime;
     }
 }
